[fix] TabBarFeature의 파일구조를 정리하고, 최소코드만 남겼습니다
