+++ Gusto-iOS/Feature/TabBarFeature/AppPath.swift
@@ -1,0 +1,6 @@
+import ComposableArchitecture
+
+@Reducer
+public enum AppPath {
+}
+
