+++ Gusto-iOS/Feature/TabBarFeature/TabEnum.swift
@@ -1,0 +1,8 @@
+
+public enum Tab: String, Equatable, CaseIterable {
+  case map
+  case review
+  case list
+  case feed
+  case my
+}
